requirements.txt

#crawling
pymongo # db driver
selenium # 크롤링 라이브러리
pandas # csv파일 생성
webdriver-manager
python-dotenv